requirements.txt
python-telegram-bot==21.6
requests
sqlalchemy
sqlalchemy-utils
pymssql
pymysql
psycopg2-binary